Move program and sub-programs into Wazup
